Allow to specify class loader(s) in EntityScanner
Note that the single-varargs method does not call
the other one because it would throw an exception
for loaders being null. I decided not to allow
null loaders because that's the same behaviour
as the loader-less method, but less clear. Readers
of the code can only assume what the first null
argument means.